backend/requirements.txt
# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
pydantic==2.5.2
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Database and Search
elasticsearch==8.11.1
redis==5.0.1

# Document Processing & OCR - Updated versions for Python 3.13
PyPDF2==3.0.1
python-docx==1.1.0
Pillow>=10.0.0
pytesseract==0.3.10
pdf2image==1.16.3

# NLP and Machine Learning
spacy==3.7.2
scikit-learn==1.3.2
numpy>=1.24.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
aiofiles==23.2.1
httpx==0.25.2
faker==20.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional utilities for production
python-dateutil==2.8.2
requests>=2.28.0